job_scraper/requirements.txt
httpx
rss-parser
thefuzz
scikit-learn
nltk
supabase
python-dotenv
openai
beautifulsoup4
lxml
